+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 * added release version info to theme.conf options section
 
 ### 0.3.2 - 2015-08-17
-* hbp-collaboratory-sphinx-theme is available as release zip asset
+* hbp-sphinx-theme is available as release zip asset